% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTIP.R
\name{CTIP}
\alias{CTIP}
\title{Computes CTIP for a given tile}
\usage{
CTIP(
  tile,
  r,
  quad.spacing,
  correction = "Ripley",
  n.null = 5,
  n.burn = 1000,
  n.sample = 11000,
  null.n.burn = 1000,
  null.n.sample = 11000,
  n.thin = 5,
  log.beta.1.mean = 0,
  log.beta.1.prec = 1e-07,
  log.beta.2.mean = 0,
  log.beta.2.prec = 1e-07,
  log.gamma.mean = 0,
  log.gamma.prec = 1e-07
)
}
\arguments{
\item{tile}{ppp object with two qualitative marks}

\item{r}{radius of interaction for model fitting.}

\item{quad.spacing}{space between points in quadrature used in estimation.}

\item{correction}{correction used in fitting; see spatstat documentation of
"ppm.ppp" for more details.}

\item{n.null}{number of null simulations; larger numbers of null simulations
improve precision, but are computationally more demanding}

\item{n.burn}{number of burn-in samples for MCMC model fitting for actual
tile as well as null tiles}

\item{n.sample}{total number of samples drawn for each simulation. The total
number of samples for the actual tile will be (n.sample - n.burn)/n.thin.
This also determines how many samples will be used to perform stochastic
integration used to estimate CTIP}

\item{n.thin}{thinning for MCMC fitting}

\item{log.beta.1.mean}{prior mean for log(beta_1), i.e. log first order
intensity of type 1 points}

\item{log.beta.1.prec}{prior precision for log(beta_1)}

\item{log.beta.2.mean}{prior mean for log(beta_2), i.e. log first order
intensity of type 2 points}

\item{log.beta.2.prec}{prior precision for log(beta_2)}

\item{log.gamma.mean}{prior mean for log(gamma), i.e. interaction parameter}

\item{log.gamma.prec}{prior precision for log(gamma)}
}
\value{
list with three attributes: "freq" gives a summary of the frequentist
fitting (for diagnostics), "model" gives the R2Jags object corresponding
to the actual fitted model, and "null.samples" gives samples from simulated
null distribution
}
\description{
Computes CTIP for a given tile
}
