% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToVis.R
\name{ExportToVis}
\alias{ExportToVis}
\title{Given an object output from the TessellateBiopsy function, formats it and
outputs it into a \code{.json} file that can be visualized using the companion
web application:}
\usage{
ExportToVis(tes, val_list, path)
}
\arguments{
\item{tes}{Output from the TessellateBiopsy function- the data you want to
visualize.}

\item{val_list}{List of values associated with each tile you'd like to
visualize. Should be a list the same length as tes$tiles, such that each
list entry has the same set of attributes each with associated values that
you'd like to visulize.}

\item{path}{Path where resulting .json file should be written to}
}
\description{
https://nateosher.github.io/SPARTIN.html
}
