% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitHSBayes.R
\name{FitHSBayes}
\alias{FitHSBayes}
\title{Fits a bayesian version of a special case of the Hierarchical Strauss Model
with gamma_{11} = gamma_{22} = 1, i.e. assuming no
interaction between points of the same type.}
\usage{
FitHSBayes(
  p,
  r,
  quad.spacing,
  correction = "Ripley",
  n.chains = 1,
  n.sample = 11000,
  n.burn = 1000,
  n.thin = 5,
  log.beta.1.mean = 0,
  log.beta.1.prec = 1e-07,
  log.beta.2.mean = 0,
  log.beta.2.prec = 1e-07,
  log.gamma.mean = 0,
  log.gamma.prec = 1e-07
)
}
\arguments{
\item{p}{spatstat ppp object with two qualitative mark levels, 1 and 2.
Note that model must have been fit using Berman-Turner Device for this to work.}

\item{r}{radius of interaction for model fitting.}

\item{quad.spacing}{space between points in quadrature used in estimation.}

\item{correction}{correction used in fitting; see spatstat documentation of
"ppm.ppp" for more details.}

\item{n.chains}{Number of chains to run}

\item{n.sample}{Total number of posterior samples}

\item{n.burn}{Number of burn-in samples}

\item{n.thin}{Thinning for chains, i.e. chains will keep sample every n.thin-th
sample from each model fitting.}

\item{log.beta.1.mean}{Mean of normal prior on log(beta_1), first order
intensity of tumor cells}

\item{log.beta.1.prec}{\emph{Precision} of normal prior on log(beta_1), the
log-first order intensity of tumor cells. Note smaller values
indicate a less informative prior.}

\item{log.beta.2.mean}{Defined analogously to log.beta.1.mean, but for
log(beta_2), the log-first order intensity of lymphocytes.}

\item{log.beta.2.prec}{Defined analogously to log.beta.1.prec, but for
log(beta_2), the log-first order intensity of lymphocytes.}

\item{log.gamma.mean}{Mean of normal prior on log(gamma_{12}), the
log-interaction parameter between tumor cells and lymphocytes. Higher values
indicate more positive interaction, lower values indicate
more negative interaction.}

\item{log.gamma.prec}{\emph{Precision} of normal prior on log(gamma_{12}), the
log-interaction parameter between tumor cells and lymphocytes. Note that
smaller values indicate a less informative prior.}
}
\value{
R2jags object
}
\description{
Fits a bayesian version of a special case of the Hierarchical Strauss Model
with gamma_{11} = gamma_{22} = 1, i.e. assuming no
interaction between points of the same type.
}
