% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitHSFreq.R
\name{FitHSFreq}
\alias{FitHSFreq}
\title{Essentially a wrapper for spatstat's ppm function; fits a
special case of the Hieararchical Strauss model that assumes
no interaction between points of the same type.}
\usage{
FitHSFreq(p, r, quad.spacing, correction = "Ripley")
}
\arguments{
\item{p}{spatstat ppp object with two qualitative mark levels, 1 and 2.}

\item{r}{radius of interaction for model fitting.}

\item{quad.spacing}{space between points in quadrature used in estimation.}

\item{correction}{correction used in fitting; see spatstat documentation of
"ppm.ppp" for more details.}
}
\value{
ppm object.
}
\description{
Essentially a wrapper for spatstat's ppm function; fits a
special case of the Hieararchical Strauss model that assumes
no interaction between points of the same type.
}
