% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTessellation.R
\name{PlotTessellation}
\alias{PlotTessellation}
\title{Plots a tessellation, colored according to the value vector passed. See
"SPARTIN-General-Usage" vignette for example.}
\usage{
PlotTessellation(tes, val, val_name = "")
}
\arguments{
\item{tes}{Tessellation from \code{TessellateBiopsy} function.}

\item{val}{Vector of values according to which the tiles should be colored.
Should be the same length as \code{tes$tiles}, i.e. one for each tile.}

\item{val_name}{String; this is what the legend will be labeled in the plot.
purely aesthetic.}
}
\value{
gg/ggplot object.
}
\description{
Plots a tessellation, colored according to the value vector passed. See
"SPARTIN-General-Usage" vignette for example.
}
