% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPPToTibble.R
\name{PPPToTibble}
\alias{PPPToTibble}
\title{Converts a marked point process (ppp) object to a tibble with
three columns: CentroidX, CentroidY, and Mark}
\usage{
PPPToTibble(p)
}
\arguments{
\item{p}{marked point process}
}
\value{
tibble with columns "CentroidX", "CentroidY", and "Mark"
}
\description{
Converts a marked point process (ppp) object to a tibble with
three columns: CentroidX, CentroidY, and Mark
}
