% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{SimulateData}
\alias{SimulateData}
\title{A basic function for simulating data}
\usage{
SimulateData(n1, n2, phi, winX, winY, r = NULL)
}
\arguments{
\item{n1}{Number of points of type 1 to be generated.}

\item{n2}{Number of points of type 2 to be generated.}

\item{phi}{Interaction parameter, a real number in \link{-1, 1}. -1 indicates most
negative Interaction; 1 indicates most positive interaction.}

\item{winX}{Horizontal width of the simulation window.}

\item{winY}{Vertical height of the simulation window.}

\item{r}{Radius of interaction; only necessary for phi > 0, i.e. positive
interaction between points.}
}
\description{
A basic function for simulating data
}
